Fix block drops & thermolith not dropping anything
Took 11 minutes

+++ src/main/java/dev/momostudios/coldsweat/common/item/SoulspringLampItem.java
@@ -68,7 +68,7 @@ public class SoulspringLampItem extends Item
                 {
                     // Drain fuel
                     if (!(player.isCreative() || player.isSpectator()))
-                        addFuel(stack, (int) (-0.01 * CSMath.clamp(temp - max, 1, 3)));
+                        addFuel(stack, -0.01 * CSMath.clamp(temp - max, 1, 3));
 
                     // Affect nearby players
                     double radius = 5d;
@@ -123,7 +123,7 @@ public class SoulspringLampItem extends Item
         stack.getOrCreateTag().putDouble("fuel", fuel);
     }
 
-    private static void addFuel(ItemStack stack, int amount)
+    private static void addFuel(ItemStack stack, double amount)
     {
         setFuel(stack, Math.min(64, getFuel(stack) + amount));
     }
